[ALPS07942820] devapc: fix build error
Fix build error caused by ACK merge

MTK-Commit-Id: 4bfe4c592e123e6b5ced667ecf8f457ab5dd0b2d

CR-Id: ALPS07942820
Feature: Kernel SI Operation

+++ drivers/soc/mediatek/devapc/devapc-mtk-multi-ao.c
@@ -1532,7 +1532,7 @@ static DRIVER_ATTR_RW(set_swp_addr);
 
 #ifdef CONFIG_MTK_SERROR_HOOK
 static void devapc_arm64_serror_panic_hook(void *data,
-		struct pt_regs *regs, unsigned int esr)
+		struct pt_regs *regs, unsigned long esr)
 {
 	struct devapc_vio_callbacks *viocb;
